feat:added register controller logic

+++ src/utils/cloudinary.js
@@ -1,5 +1,8 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
+// import dotenv from "dotenv";
+
+// dotenv.config();
 
 
 cloudinary.config({ 
@@ -12,18 +15,26 @@ const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null
         //upload the file on cloudinary
-        const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
+        const response = await cloudinary.uploader.upload(
+  localFilePath.replace(/\\/g, "/"),
+  { resource_type: "auto", timeout: 60000 }   // 60s
+);
+
         // file has been uploaded successful
         //console.log("file is uploaded on cloudinary ", response.url);
         fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
-    }
+  console.log("Cloudinary upload error --->", error?.message || error);
+
+  if (localFilePath && fs.existsSync(localFilePath)) {
+    fs.unlinkSync(localFilePath);
+  }
+
+  return null;
+}
+
 }
